+++ src/Http/Controllers/V1/Admin/Customers/CustomerBaseController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Webkul\RestApi\Http\Controllers\V1\Admin\Customer;
+namespace Webkul\RestApi\Http\Controllers\V1\Admin\Customers;
 
 use Webkul\RestApi\Http\Controllers\V1\Admin\AdminController;
