+++ src/Http/Controllers/V1/Admin/Marketing/Communications/TemplateController.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace Webkul\RestApi\Http\Controllers\V1\Admin\Marketing;
+namespace Webkul\RestApi\Http\Controllers\V1\Admin\Marketing\Communications;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Event;
 use Webkul\Marketing\Repositories\TemplateRepository;
 use Webkul\RestApi\Http\Resources\V1\Admin\Marketing\TemplateResource;
+use Webkul\RestApi\Http\Controllers\V1\Admin\Marketing\MarketingController;
 
 class TemplateController extends MarketingController
 {
