+++ src/Http/Controllers/V1/Admin/Settings/InventorySourceController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Webkul\RestApi\Http\Controllers\V1\Admin\Setting;
+namespace Webkul\RestApi\Http\Controllers\V1\Admin\Settings;
 
 use Illuminate\Support\Facades\Event;
 use Webkul\Admin\Http\Requests\InventorySourceRequest;
