+++ src/Http/Controllers/V1/Admin/Settings/Tax/TaxCategoryController.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace Webkul\RestApi\Http\Controllers\V1\Admin\Setting;
+namespace Webkul\RestApi\Http\Controllers\V1\Admin\Settings\Tax;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Event;
 use Webkul\RestApi\Http\Resources\V1\Admin\Tax\TaxCategoryResource;
 use Webkul\Tax\Repositories\TaxCategoryRepository;
+use Webkul\RestApi\Http\Controllers\V1\Admin\Settings\SettingController;
 
 class TaxCategoryController extends SettingController
 {
